now multi-threaded to support multiple api keys

+++ download.py
@@ -1,10 +1,15 @@
+import random
+import string
 import requests
+import threading
 import json
 import os
 from os import path
 from time import sleep
 
-api_key = open('eddy.apikey').read().strip()
+api_keys = [open('eddy.apikey').read().strip(),
+            open('jason.apikey').read().strip()]
+api_key = api_keys[0] 
 input_directory = path.join('AP_ITEM_DATASET','5.11','RANKED_SOLO')
 output = 'output'
 output_directory = path.join(input_directory, output) 
@@ -32,29 +37,55 @@ if not path.exists(output_directory):
     print "creating output directory %s" % output_directory
     os.mkdir(output_directory)
 
+def process(api_key, region, name=-1):
+    global games
+    while (games):
+        game = games.pop()
+        size = len(games)
+        filename = path.join(output_directory, region, str(game) + '.json')
+        if not path.isfile(filename):
+            print "[Thread %d] Downloading %s (%d left)" % (name, filename, size)
+            ok = False
+            while not ok:
+                r = requests.get(apiBaseUrl[region.lower()] + match_api % (region.lower(), game), params={'api_key': api_key, 'includeTimeline': 'true'})
+                if r.status_code == 429:
+                    print "Retrying in 10 seconds"
+                    sleep(10)
+                elif r.status_code != 200:
+                    print 'received bad status code %s for game %s, retrying in a minute' % (game, r.status_code)
+                    sleep(60)
+                    #raise RuntimeError('returned bad status code %s' % r.status_code)
+                else:
+                    ok = True
+            tempname = ''.join(random.choice(string.ascii_uppercase + string.digits) for _ in range(8)) 
+            outfile = open(tempname, 'w')
+            outfile.write(r.text.encode('utf-8'))
+            # outfile.close()
+            try:
+                os.rename(tempname,filename)
+            except Exception as e:
+                if path.isfile(tempname): os.remove(tempname)
+                if path.isfile(filename): os.remove(filename)
+                raise e
+            sleep(1.5)
 
 for f in [f for f in os.listdir(input_directory) if path.isfile(path.join(input_directory, f))]:
     region = f[:-5]
-    if region != 'NA': continue #only NA supported for now
+    #if region != 'NA': continue #only NA supported for now
     if not path.exists(path.join(output_directory, region)):
         os.mkdir(path.join(output_directory, region))
         print "creating output directory %s" % path.join(output_directory, region)
-    games = json.load(open(path.join(input_directory,f)))
-    for i,game in enumerate(games):
-        filename = path.join(output_directory, region, str(game) + '.json')
-        if not path.isfile(filename):
-            print "Downloading %s (%d of %d)" % (filename, i+1, len(games))
-            r = requests.get(apiBaseUrl[region.lower()] + match_api % (region.lower(), game), params={'api_key': api_key, 'includeTimeline': 'true'})
-            while r.status_code == 429:
-                print "Retrying in 10 seconds"
-                sleep(10)
-                r = requests.get(apiBaseUrl[region.lower()] + match_api % (region.lower(), game), params={'api_key': api_key, 'includeTimeline': 'true'})
-            outfile = open('temp', 'w')
-            outfile.write(r.text.encode('utf-8'))
-            outfile.close()
-            try:
-                os.rename('temp',filename)
-            except:
-                os.remove('temp')
-                os.remove(filename)
-            sleep(1.5)
+    games = set([game for game in json.load(open(path.join(input_directory,f))) if not path.isfile(path.join(output_directory, region, str(game) + '.json'))])
+    threads = []
+    q = 0
+    for apikey in api_keys:
+        threads.append(threading.Thread(target=process, args=[apikey, region, q]))
+        q += 1
+    for thread in threads:
+        thread.daemon = True
+        thread.start()
+    while threading.active_count():
+        pass
+    # while all([thread.is_alive() for thread in threads]):
+    #     pass
+
